Clean up syntax error types. Move to own file.
