+++ packages/runty/src/index.ts
@@ -1,9 +1,9 @@
-export { fns } from './allFns';
-export { RSyntaxError } from './constants';
+export * from './allFns';
 export * from './fns';
-export { logicFns } from './logicFns';
-export { numberFns } from './numberFns';
-export { array, runty, string } from './runty';
-export { stringFns } from './stringFns';
+export * from './logicFns';
+export * from './numberFns';
+export * from './RSyntaxError';
+export * from './runty';
+export * from './stringFns';
 export * from './types';
-export { utilFns } from './utilFns';
+export * from './utilFns';
